exp20(harness): DocStreamer uint16 shard format + multi-shard + EOS-less tests
Code review caught a plan defect — spec said np.int32 but canonical SP shards
(src/chaoscontrol/data.py:79, scripts/build_sp_shards.py:54) are uint16. Int32
path would silently misread real shards. Also adds two missing edge-case tests
(multi-shard continuous doc_id, shard without trailing EOS).

+++ src/chaoscontrol/eval_stream/doc_stream.py
@@ -29,7 +29,7 @@ class DocStreamer:
     def __iter__(self) -> Iterator[DocRecord]:
         doc_id = 0
         for shard in self.shard_paths:
-            arr = np.fromfile(str(shard), dtype=np.int32)
+            arr = np.fromfile(str(shard), dtype=np.uint16)
             buf: list[int] = []
             for t in arr:
                 t_int = int(t)
